+++ CHANGELOG.rst
@@ -19,17 +19,33 @@ Added
 - ``lensless.hardware.trainable_mask.TrainableCodedAperture`` class for training a coded aperture mask pattern.
 - Support for other optimizers in ``lensless.utils.Trainer.set_optimizer``.
 - ``lensless.utils.dataset.simulate_dataset`` for simulating a dataset given a mask/PSF.
+- Support for training/testing with multiple mask patterns in the dataset.
+- Multi-GPU support for training.
+- DigiCam dataset which interfaces with Hugging Face.
+- Scripts for authentication.
+- DigiCam support for Telegram demo.
+- DiffuserCamMirflickr Hugging Face API.
+- Fallback for normalization if data not in 8bit range (``lensless.utils.io.save_image``).
+- Add utilities for fabricating masks with 3D printing (``lensless.hardware.fabrication``).
 
 Changed
-~~~~~
+~~~~~~~
 
 - Dataset reconstruction script uses datasets from Hugging Face: ``scripts/recon/dataset.py``
 - For trainable masks, set trainable parameters inside the child class.
+- ``distance_sensor`` optional for ``lensless.hardware.mask.Mask``, e.g. don't need for fabrication.
+- More intuitive interface for MURA for coded aperture (``lensless.hardware.mask.CodedAperture``), i.e. directly pass prime number.
+
 
 Bugfix
-~~~~~
+~~~~~~
 
 - ``lensless.hardware.trainable_mask.AdafruitLCD`` input handling.
+- Local path for DRUNet download.
+- APGD input handling (float32).
+- Multimask handling.
+- Passing shape to IRFFT so that it matches shape of input to RFFT.
+- MLS mask creation (needed to rescale digits).
 
 1.0.6 - (2024-02-21)
 --------------------